apps/goldtrack/requirements.txt
flask
psycopg2-binary
yfinance
requests
pyjwt
gunicorn